Move pcluster delete code to self contained python module
